requirements.txt
albumentations==0.5.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
bokeh==2.3.1
certifi==2020.12.5
cffi==1.14.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
efficientnet-pytorch==0.6.3
entrypoints==0.3
h5py==3.2.1
imageio==2.9.0
imgaug==0.4.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
livelossplot==0.5.4
MarkupSafe==1.1.1
mat73==0.46
matplotlib==3.3.3
MedPy==0.4.0
mistune==0.8.4
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.3.0
numpy==1.19.5
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plyfile==0.7.3
pretrainedmodels==0.7.4
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
Shapely==1.7.1
SimpleITK==2.0.2
six==1.15.0
sklearn==0.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.14
timm==0.3.2
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
